+++ src/commands/connect.c
@@ -49,6 +49,11 @@ static bool quit_reconnecting = false;
 static bool reconnecting = false;
 static bool secure_connection = false;
 
+static const char *anonops_servers[] = {
+    "anonops532vcpz6z.onion",
+    "irc.anonops.com",
+};
+
 static const char *efnet_servers[] = {
     "efnet.port80.se",
     "efnet.portlane.se",
@@ -404,7 +409,10 @@ cmd_connect(const char *data)
 	print_and_free("/connect: bogus port number", dcopy);
 	return;
     } else {
-	if (strings_match_ignore_case(server, "efnet"))
+	if (strings_match_ignore_case(server, "anonops"))
+	    IRC_CONNECT(get_server(anonops_servers,
+		ARRAY_SIZE(anonops_servers), "AnonOps IRC network"), port);
+	else if (strings_match_ignore_case(server, "efnet"))
 	    IRC_CONNECT(get_server(efnet_servers,
 		ARRAY_SIZE(efnet_servers), "EFnet servers"), port);
 	else if (strings_match_ignore_case(server, "freenode"))